Fix undefined/unclear behavior, and clean up code

+++ Boot/NewLegacy/Init/Disk/Disk.h
@@ -12,8 +12,8 @@
   extern uint16 LogicalSectorSize;
   extern uint16 PhysicalSectorSize;
 
-  realModeTable* ReadSector(uint16 NumBlocks, uint64 Address, uint64 Offset);
-  realModeTable* ReadLogicalSector(uint16 NumBlocks, uint64 Address, uint32 Lba);
+  realModeTable* ReadSector(uint16 NumBlocks, uint32 Address, uint64 Offset);
+  realModeTable* ReadLogicalSector(uint16 NumBlocks, uint32 Address, uint32 Lba);
 
   // Filesystem-related functions. (Disk.c)
